requirements.txt
python-telegram-bot
requests
mercadopago
flask
